move logger
