+++ go/pkg/interface/handler/account_handler.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 
 	"l-semi-chat/pkg/domain"
-	"l-semi-chat/pkg/interface/server/logger"
+	"l-semi-chat/pkg/domain/logger"
 	"l-semi-chat/pkg/interface/server/response"
 	"l-semi-chat/pkg/service/interactor"
